Broken encoder and decoder with instances
Note : This commit is a checkpoint were I realised that my encoding and
decoding was going in the wrong direction.

It adds the FieldWithMedata type representing a single field's data with
some associated metadata along with its Circe instances. The SchemaTree
type is also introduced for holding together the fields. However, its
implementation is flawed and the circe instances are broken.