Remove SchemaTree instances manual encode/decode
This commit removes the instances of Foldable and Functor from
SchemaTree[T]. The reasoning behind this is that they were unusable in
the context of the decoder because it was necessary to get the next
entity's id to generate the JSON block.

Another solution could have been to use a Zipper. However, I wanted to
keep the code as simple and flexible as possible.