Add Response data structure and validation
This commit adds the response data type and a validation function
signature. It also introduces the Zipper data structure see :
https://en.wikipedia.org/wiki/Zipper_(data_structure). This will allow
easier traversal of the SchemaTree while retaining the context.
I believe it was slightly over engineered, so this commit also serves as
a checkpoint for that.

Other changes in random order :

 * Fix compilation failure due to CreateVersionRequest suppression

 * Arrange some imports