Attempt at plugins through interface approach
What I tried is to make a list of plugin traits. With the idea that
anyone trying to implement a plugin will implement the trait, and I will
just load all of them in a list before using them.

Oh how wrong I was. The issue is that Decoder and Encoder require an
invariant type argument. I guess it makes sense, but then for it to compile,
the plugin signature ends up being Plugin[A <: Field, B <: Answer] and not
Plugin[+A <: Field, +B <: Answer], preventing me to store a
Plugin[field_.Text, answer_.Text] in a Plugin[field_.Field,
answer_.Answer]. So when calling decode I cannot get back a field or an
answer.

If in the future anyone reads this and doubts, please just checkout
Plugins.scala and try to compile it to do decoding.